refact: 코드 구조 변경, 디자인 일부 수정
